Re factor: Move Profile Response
